+++ cpp/DumbHostObject.cpp
@@ -1,5 +1,5 @@
-#include "DumbHostObject.h"
-#include "SmartHostObject.h"
+#include "DumbHostObject.hpp"
+#include "SmartHostObject.hpp"
 #include "utils.hpp"
 #include <iostream>
