+++ cpp/SmartHostObject.hpp
